requirements.txt
beautifulsoup4==4.12.3
langchain==0.2.7
langchain-core==0.2.7
langgraph==0.1.5
langchainhub==0.1.20
langchain-community==0.2.7
tavily-python==0.3.4
faiss-cpu==1.7.4
python-dotenv==1.0.1
pytest==8.2.2
langchain-google-genai==0.0.6
google-generativeai==0.3.2
tiktoken==0.5.2
